src/semantic_understanding/multilingual_semantic_engine.py
import torch
from transformers import (
    AutoTokenizer,
    AutoModel,
    AutoModelForSequenceClassification,
    AutoModelForTokenClassification
)
from typing import Dict, List, Any, Optional, Tuple, Set
import numpy as np
from dataclasses import dataclass, field
from sklearn.metrics.pairwise import cosine_similarity
import networkx as nx

@dataclass
class SemanticRepresentation:
    """
    Represents multilingual semantic understanding
    """
    text: str
    language: str
    embedding: torch.Tensor
    concepts: List[str]
    relations: List[Tuple[str, str, str]]
    confidence: float
    metadata: Dict[str, Any] = field(default_factory=dict)

@dataclass
class CrossLingualAlignment:
    """
    Represents alignment between different languages
    """
    source_lang: str
    target_lang: str
    alignment_score: float
    translation_pairs: List[Tuple[str, str]]
    semantic_drift: float
    confidence: float

class MultilingualSemanticEngine:
    def __init__(
        self,
        supported_languages: Optional[List[str]] = None,
        model_config: Optional[Dict[str, Any]] = None
    ):
        """
        Initialize Multilingual Semantic Engine
        
        Args:
            supported_languages: List of supported language codes
            model_config: Model configuration parameters
        """
        self.supported_languages = supported_languages or [
            'en', 'es', 'fr', 'de', 'zh', 'ja', 'ko', 'ar', 'ru'
        ]
        
        # Initialize configuration
        self.config = model_config or {
            'base_model': 'xlm-roberta-large',
            'max_sequence_length': 512,
            'semantic_threshold': 0.75,
            'alignment_threshold': 0.8,
            'batch_size': 32
        }
        
        # Initialize models
        self.tokenizer = AutoTokenizer.from_pretrained(self.config['base_model'])
        self.encoder = AutoModel.from_pretrained(self.config['base_model'])
        
        # Task-specific models
        self.models = {
            'classification': AutoModelForSequenceClassification.from_pretrained(self.config['base_model']),
            'ner': AutoModelForTokenClassification.from_pretrained(self.config['base_model'])
        }
        
        # Language-specific analyzers
        self.language_analyzers = self._initialize_language_analyzers()
        
        # Semantic graph for cross-lingual relationships
        self.semantic_graph = nx.MultiDiGraph()
    
    def analyze_text(
        self,
        text: str,
        source_lang: str,
        target_langs: Optional[List[str]] = None
    ) -> SemanticRepresentation:
        """
        Analyze text and generate semantic representation
        
        Args:
            text: Input text
            source_lang: Source language code
            target_langs: Target language codes for cross-lingual analysis
        
        Returns:
            Semantic representation
        """
        # Validate language support
        if source_lang not in self.supported_languages:
            raise ValueError(f"Unsupported language: {source_lang}")
        
        # Encode text
        encoding = self._encode_text(text, source_lang)
        
        # Extract concepts and relations
        concepts = self._extract_concepts(encoding, source_lang)
        relations = self._extract_relations(encoding, concepts)
        
        # Create semantic representation
        representation = SemanticRepresentation(
            text=text,
            language=source_lang,
            embedding=encoding['embedding'],
            concepts=concepts,
            relations=relations,
            confidence=encoding['confidence']
        )
        
        # Perform cross-lingual analysis if requested
        if target_langs:
            representation.metadata['cross_lingual'] = self._analyze_cross_lingual(
                representation,
                target_langs
            )
        
        return representation
    
    def align_languages(
        self,
        source_texts: List[str],
        target_texts: List[str],
        source_lang: str,
        target_lang: str
    ) -> CrossLingualAlignment:
        """
        Align semantic representations across languages
        
        Args:
            source_texts: Source language texts
            target_texts: Target language texts
            source_lang: Source language code
            target_lang: Target language code
        
        Returns:
            Cross-lingual alignment
        """
        # Validate languages
        for lang in [source_lang, target_lang]:
            if lang not in self.supported_languages:
                raise ValueError(f"Unsupported language: {lang}")
        
        # Generate embeddings
        source_embeddings = self._batch_encode_texts(source_texts, source_lang)
        target_embeddings = self._batch_encode_texts(target_texts, target_lang)
        
        # Compute alignment score
        alignment_score = self._compute_alignment_score(
            source_embeddings,
            target_embeddings
        )
        
        # Find translation pairs
        translation_pairs = self._find_translation_pairs(
            source_texts,
            target_texts,
            source_embeddings,
            target_embeddings
        )
        
        # Compute semantic drift
        semantic_drift = self._compute_semantic_drift(
            source_embeddings,
            target_embeddings
        )
        
        return CrossLingualAlignment(
            source_lang=source_lang,
            target_lang=target_lang,
            alignment_score=alignment_score,
            translation_pairs=translation_pairs,
            semantic_drift=semantic_drift,
            confidence=self._compute_alignment_confidence(alignment_score, semantic_drift)
        )
    
    def transfer_knowledge(
        self,
        source_representation: SemanticRepresentation,
        target_lang: str
    ) -> SemanticRepresentation:
        """
        Transfer semantic knowledge to target language
        
        Args:
            source_representation: Source semantic representation
            target_lang: Target language code
        
        Returns:
            Transferred semantic representation
        """
        # Validate target language
        if target_lang not in self.supported_languages:
            raise ValueError(f"Unsupported language: {target_lang}")
        
        # Transfer concepts and relations
        transferred_concepts = self._transfer_concepts(
            source_representation.concepts,
            source_representation.language,
            target_lang
        )
        
        transferred_relations = self._transfer_relations(
            source_representation.relations,
            source_representation.language,
            target_lang
        )
        
        # Generate target language embedding
        target_embedding = self._transfer_embedding(
            source_representation.embedding,
            source_representation.language,
            target_lang
        )
        
        return SemanticRepresentation(
            text="",  # No direct text translation
            language=target_lang,
            embedding=target_embedding,
            concepts=transferred_concepts,
            relations=transferred_relations,
            confidence=source_representation.confidence * 0.9  # Slight confidence reduction
        )
    
    def _encode_text(
        self,
        text: str,
        language: str
    ) -> Dict[str, Any]:
        """
        Encode text into semantic representation
        
        Args:
            text: Input text
            language: Language code
        
        Returns:
            Encoding information
        """
        # Tokenize text
        inputs = self.tokenizer(
            text,
            return_tensors='pt',
            max_length=self.config['max_sequence_length'],
            truncation=True,
            padding=True
        )
        
        # Generate embeddings
        with torch.no_grad():
            outputs = self.encoder(**inputs)
            embedding = outputs.last_hidden_state.mean(dim=1)
        
        return {
            'embedding': embedding,
            'attention_weights': outputs.attentions,
            'confidence': self._compute_encoding_confidence(outputs)
        }
    
    def _batch_encode_texts(
        self,
        texts: List[str],
        language: str
    ) -> torch.Tensor:
        """
        Encode multiple texts in batches
        
        Args:
            texts: List of input texts
            language: Language code
        
        Returns:
            Batch of embeddings
        """
        embeddings = []
        
        for i in range(0, len(texts), self.config['batch_size']):
            batch_texts = texts[i:i + self.config['batch_size']]
            
            # Tokenize batch
            inputs = self.tokenizer(
                batch_texts,
                return_tensors='pt',
                max_length=self.config['max_sequence_length'],
                truncation=True,
                padding=True
            )
            
            # Generate embeddings
            with torch.no_grad():
                outputs = self.encoder(**inputs)
                batch_embeddings = outputs.last_hidden_state.mean(dim=1)
                embeddings.append(batch_embeddings)
        
        return torch.cat(embeddings, dim=0)
    
    def _extract_concepts(
        self,
        encoding: Dict[str, Any],
        language: str
    ) -> List[str]:
        """
        Extract concepts from encoded representation
        
        Args:
            encoding: Encoded text information
            language: Language code
        
        Returns:
            List of extracted concepts
        """
        # Use language-specific analyzer
        analyzer = self.language_analyzers.get(language)
        if analyzer:
            return analyzer.extract_concepts(encoding)
        
        # Fallback to generic concept extraction
        return self._generic_concept_extraction(encoding)
    
    def _extract_relations(
        self,
        encoding: Dict[str, Any],
        concepts: List[str]
    ) -> List[Tuple[str, str, str]]:
        """
        Extract relations between concepts
        
        Args:
            encoding: Encoded text information
            concepts: Extracted concepts
        
        Returns:
            List of relations (subject, predicate, object)
        """
        relations = []
        
        # Use attention weights to identify relationships
        attention_weights = encoding['attention_weights']
        
        # Process each layer of attention
        for layer_attention in attention_weights:
            # Analyze attention patterns
            relations.extend(
                self._analyze_attention_patterns(layer_attention, concepts)
            )
        
        return self._filter_relations(relations)
    
    def _analyze_cross_lingual(
        self,
        representation: SemanticRepresentation,
        target_langs: List[str]
    ) -> Dict[str, Any]:
        """
        Perform cross-lingual analysis
        
        Args:
            representation: Source semantic representation
            target_langs: Target language codes
        
        Returns:
            Cross-lingual analysis results
        """
        results = {}
        
        for target_lang in target_langs:
            if target_lang != representation.language:
                # Transfer knowledge to target language
                transferred = self.transfer_knowledge(representation, target_lang)
                
                # Compute alignment
                alignment = self._compute_language_alignment(
                    representation,
                    transferred
                )
                
                results[target_lang] = {
                    'alignment_score': alignment.alignment_score,
                    'semantic_drift': alignment.semantic_drift,
                    'confidence': alignment.confidence
                }
        
        return results
    
    def _compute_alignment_score(
        self,
        source_embeddings: torch.Tensor,
        target_embeddings: torch.Tensor
    ) -> float:
        """
        Compute alignment score between language embeddings
        
        Args:
            source_embeddings: Source language embeddings
            target_embeddings: Target language embeddings
        
        Returns:
            Alignment score
        """
        # Convert to numpy for similarity computation
        source_np = source_embeddings.numpy()
        target_np = target_embeddings.numpy()
        
        # Compute cosine similarity
        similarity_matrix = cosine_similarity(source_np, target_np)
        
        return float(similarity_matrix.mean())
    
    def _find_translation_pairs(
        self,
        source_texts: List[str],
        target_texts: List[str],
        source_embeddings: torch.Tensor,
        target_embeddings: torch.Tensor
    ) -> List[Tuple[str, str]]:
        """
        Find translation pairs based on semantic similarity
        
        Args:
            source_texts: Source language texts
            target_texts: Target language texts
            source_embeddings: Source language embeddings
            target_embeddings: Target language embeddings
        
        Returns:
            List of translation pairs
        """
        pairs = []
        
        # Compute similarity matrix
        similarity_matrix = cosine_similarity(
            source_embeddings.numpy(),
            target_embeddings.numpy()
        )
        
        # Find best matches
        for i in range(len(source_texts)):
            best_match = similarity_matrix[i].argmax()
            if similarity_matrix[i][best_match] >= self.config['alignment_threshold']:
                pairs.append((source_texts[i], target_texts[best_match]))
        
        return pairs
    
    def _compute_semantic_drift(
        self,
        source_embeddings: torch.Tensor,
        target_embeddings: torch.Tensor
    ) -> float:
        """
        Compute semantic drift between languages
        
        Args:
            source_embeddings: Source language embeddings
            target_embeddings: Target language embeddings
        
        Returns:
            Semantic drift score
        """
        # Convert to numpy
        source_np = source_embeddings.numpy()
        target_np = target_embeddings.numpy()
        
        # Compute distribution difference
        source_mean = source_np.mean(axis=0)
        target_mean = target_np.mean(axis=0)
        
        return float(np.linalg.norm(source_mean - target_mean))
    
    def _transfer_concepts(
        self,
        concepts: List[str],
        source_lang: str,
        target_lang: str
    ) -> List[str]:
        """
        Transfer concepts to target language
        
        Args:
            concepts: Source language concepts
            source_lang: Source language code
            target_lang: Target language code
        
        Returns:
            Transferred concepts
        """
        transferred = []
        
        for concept in concepts:
            # Look up concept in semantic graph
            if self.semantic_graph.has_node(concept):
                # Find target language equivalent
                for _, target_node, data in self.semantic_graph.edges(concept, data=True):
                    if data['language'] == target_lang:
                        transferred.append(target_node)
                        break
        
        return transferred
    
    def _transfer_relations(
        self,
        relations: List[Tuple[str, str, str]],
        source_lang: str,
        target_lang: str
    ) -> List[Tuple[str, str, str]]:
        """
        Transfer relations to target language
        
        Args:
            relations: Source language relations
            source_lang: Source language code
            target_lang: Target language code
        
        Returns:
            Transferred relations
        """
        transferred = []
        
        for subject, predicate, obj in relations:
            # Transfer each component
            transferred_subject = self._transfer_concepts([subject], source_lang, target_lang)[0]
            transferred_predicate = self._transfer_concepts([predicate], source_lang, target_lang)[0]
            transferred_object = self._transfer_concepts([obj], source_lang, target_lang)[0]
            
            transferred.append((
                transferred_subject,
                transferred_predicate,
                transferred_object
            ))
        
        return transferred
    
    def _transfer_embedding(
        self,
        embedding: torch.Tensor,
        source_lang: str,
        target_lang: str
    ) -> torch.Tensor:
        """
        Transfer embedding to target language space
        
        Args:
            embedding: Source language embedding
            source_lang: Source language code
            target_lang: Target language code
        
        Returns:
            Transferred embedding
        """
        # Apply language-specific transformation
        transformation_matrix = self._get_language_transformation(
            source_lang,
            target_lang
        )
        
        return torch.matmul(embedding, transformation_matrix)
    
    def _initialize_language_analyzers(self) -> Dict[str, Any]:
        """
        Initialize language-specific analyzers
        
        Returns:
            Dictionary of language analyzers
        """
        # Placeholder for language-specific analyzers
        return {
            'en': self._create_english_analyzer(),
            'es': self._create_spanish_analyzer(),
            'fr': self._create_french_analyzer(),
            'de': self._create_german_analyzer(),
            'zh': self._create_chinese_analyzer(),
            'ja': self._create_japanese_analyzer(),
            'ko': self._create_korean_analyzer(),
            'ar': self._create_arabic_analyzer(),
            'ru': self._create_russian_analyzer()
        }
    
    # Placeholder methods for language-specific analyzers
    def _create_english_analyzer(self):
        return None  # Placeholder
    
    def _create_spanish_analyzer(self):
        return None  # Placeholder
    
    def _create_french_analyzer(self):
        return None  # Placeholder
    
    def _create_german_analyzer(self):
        return None  # Placeholder
    
    def _create_chinese_analyzer(self):
        return None  # Placeholder
    
    def _create_japanese_analyzer(self):
        return None  # Placeholder
    
    def _create_korean_analyzer(self):
        return None  # Placeholder
    
    def _create_arabic_analyzer(self):
        return None  # Placeholder
    
    def _create_russian_analyzer(self):
        return None  # Placeholder
